+++ plugins/contact-resources/src/components/EmployeeArrayEditor.svelte
@@ -1,17 +1,16 @@
 <script lang="ts">
-  import { Person } from '@anticrm/contact'
+  import { Employee } from '@anticrm/contact'
   import { Ref } from '@anticrm/core'
   import { IntlString } from '@anticrm/platform'
   import { UserBoxList } from '@anticrm/presentation'
-  import contact from '../plugin'
 
   export let label: IntlString
-  export let value: Ref<Person>[]
-  export let onChange: (refs: Ref<Person>[]) => void
+  export let value: Ref<Employee>[]
+  export let onChange: (refs: Ref<Employee>[]) => void
 
   let timer: any
 
-  function onUpdate (evt: CustomEvent<Ref<Person>[]>): void {
+  function onUpdate (evt: CustomEvent<Ref<Employee>[]>): void {
     clearTimeout(timer)
     timer = setTimeout(() => {
       onChange(evt.detail)
@@ -19,13 +18,4 @@
   }
 </script>
 
-<UserBoxList
-  _class={contact.class.Employee}
-  items={value}
-  {label}
-  on:update={onUpdate}
-  kind={'link'}
-  size={'medium'}
-  justify={'left'}
-  width={'100%'}
-/>
+<UserBoxList items={value} {label} on:update={onUpdate} kind={'link'} size={'medium'} justify={'left'} width={'100%'} />