+++ plugins/tracker-resources/src/components/projects/LeadPresenter.svelte
@@ -72,6 +72,9 @@
       {
         _class: contact.class.Employee,
         selected: value?._id,
+        docQuery: {
+          active: true
+        },
         allowDeselect: true,
         placeholder: tracker.string.ProjectLeadSearchPlaceholder
       },